Add files via upload
The Resume-Portfolio Builder is a dynamic application developed using Flutter, Dart, HTML, CSS, and JavaScript. It allows users to create professional resumes and portfolios effortlessly through a user-friendly interface. The application offers customizable templates that cater to various professional needs, enabling users to personalize their documents to reflect their unique styles and preferences. Flutter and Dart provide a seamless and responsive user experience, while HTML, CSS, and JavaScript enhance the design and interactivity of the application. This tool is ideal for individuals looking to showcase their skills and achievements in a polished and professional manner.